feat: market headers, reformatted lp page
