--- src/pages/liquidity/[...id].tsx
@@ -1,293 +1,0 @@
-import React, { useEffect, useState, useCallback } from 'react'
-import styled from 'styled-components'
-import PageHeader from '@/components/PageHeader'
-import MarketCards from '@/components/MarketCards'
-import Spacer from '@/components/Spacer'
-import Box from '@/components/Box'
-import Loader from '@/components/Loader'
-import Disclaimer from '@/components/Disclaimer'
-
-import MetaMaskOnboarding from '@metamask/onboarding'
-
-import Notifs from '@/components/Notifs'
-import Tooltip from '@/components/Tooltip'
-
-import { useActiveWeb3React } from '@/hooks/user/index'
-import ErrorBoundary from '@/components/ErrorBoundary'
-import { useRouter } from 'next/router'
-import { GetServerSideProps } from 'next'
-import { useOptions, useUpdateOptions } from '@/state/options/hooks'
-import { useClearNotif } from '@/state/notifs/hooks'
-import { useSetLoading } from '@/state/positions/hooks'
-
-import { Grid, Col, Row } from 'react-styled-flexboxgrid'
-
-import LiquidityTable from '@/components/Liquidity/LiquidityTable'
-import numeral from 'numeral'
-import { formatEther } from 'ethers/lib/utils'
-import isZero from '@/utils/isZero'
-
-import PositionsCard from '@/components/Market/PositionsCard'
-import Switch from '@/components/Switch'
-import MarketHeader from '@/components/Market/MarketHeader'
-import FilterBar from '@/components/Market/FilterBar'
-import { useRemoveItem } from '@/state/order/hooks'
-
-export const getServerSideProps: GetServerSideProps = async ({ params }) => {
-  const data = params?.id
-
-  return {
-    props: {
-      user: data[0],
-      data: data,
-    },
-  }
-}
-
-interface CardProps {
-  title?: string
-  description?: string
-  multiplier?: number
-}
-
-export const DataCard: React.FC<CardProps> = ({
-  children,
-  title,
-  description,
-  multiplier,
-}) => {
-  return (
-    <StyledCardContainer multiplier={multiplier ? multiplier : 2}>
-      <CardContent>
-        <CardTitle>{title ? title : 'No title'}</CardTitle>
-      </CardContent>
-      <CardContent>
-        <Text>
-          {description ? description : children ? children : 'No description'}
-        </Text>
-      </CardContent>
-    </StyledCardContainer>
-  )
-}
-
-const StyledL = styled(Box)`
-  margin-top: -1.5em;
-  margin-bottom: -1.1em;
-`
-
-export const Graph: React.FC = () => {
-  return <Box>Graph</Box>
-}
-
-const Liquidity = ({ user, data }) => {
-  const [callPutActive, setCallPutActive] = useState(true)
-  const { chainId, active, account, library } = useActiveWeb3React()
-  const [id, storeId] = useState(chainId)
-  const [changing, setChanging] = useState(false)
-  const router = useRouter()
-  const clear = useClearNotif()
-  const options = useOptions()
-  const updateOptions = useUpdateOptions()
-  const setLoading = useSetLoading()
-  const removeItem = useRemoveItem()
-
-  useEffect(() => {
-    updateOptions('weth') //hardcoded
-  }, [user])
-
-  useEffect(() => {
-    const { ethereum, web3 } = window as any
-
-    if (MetaMaskOnboarding.isMetaMaskInstalled() && (!ethereum || !web3)) {
-      clear(0)
-      router.push(`/liquidity`)
-    }
-    if (ethereum) {
-      const handleChainChanged = () => {
-        if (id !== chainId) {
-          setChanging(true)
-          storeId(chainId)
-          // eat errors
-          clear(0)
-          router.reload()
-        }
-      }
-      const handleAccountChanged = () => {
-        if (!options.loading) {
-          clear(0)
-          setLoading()
-        }
-      }
-      if (ethereum?.on) {
-        ethereum?.on('chainChanged', handleChainChanged)
-        ethereum?.on('accountsChanged', handleAccountChanged)
-      }
-      return () => {
-        if (ethereum?.removeListener) {
-          ethereum.removeListener('chainChanged', handleChainChanged)
-          ethereum.removeListener('accountsChanged', handleAccountChanged)
-        }
-      }
-    }
-  }, [id, chainId, storeId])
-
-  if (!active) {
-    return (
-      <>
-        <Spacer />
-        <Loader size="lg" />
-      </>
-    )
-  }
-  if (!(chainId === 4 || chainId === 1) && active) {
-    return <Text>Please switch to Rinkeby or Mainnet Networks</Text>
-  }
-  if (
-    !MetaMaskOnboarding.isMetaMaskInstalled() ||
-    !(window as any)?.ethereum ||
-    !(window as any)?.web3
-  ) {
-    return <Text>Please Install Metamask to View Markets</Text>
-  }
-  if (MetaMaskOnboarding.isMetaMaskInstalled() && !account) {
-    return <Text>Please Connect to Metamask to View Markets</Text>
-  }
-
-  return (
-    <ErrorBoundary
-      fallback={
-        <>
-          <Spacer />
-          <Text>Error Loading Liquidity, Please Refresh</Text>
-        </>
-      }
-    >
-      {changing ? (
-        <>
-          <Spacer />
-          <Spacer />
-          <Loader size="lg" />
-        </>
-      ) : (
-        <>
-          <Disclaimer />
-          <StyledMarket>
-            <Grid id={'market-grid'}>
-              <Row>
-                <StyledLitContainer>
-                  <div>
-                    <StyledHeaderContainer>
-                      <MarketHeader
-                        marketId={'weth'}
-                        isCall={callPutActive ? 0 : 1}
-                      />
-                      <FilterBar
-                        active={callPutActive}
-                        setCallActive={() => setCallPutActive(!callPutActive)}
-                      />
-                    </StyledHeaderContainer>
-
-                    <StyledHeaderContainer>
-                      <LiquidityTable callActive={callPutActive} />
-                    </StyledHeaderContainer>
-                  </div>
-                </StyledLitContainer>
-              </Row>
-            </Grid>
-          </StyledMarket>
-        </>
-      )}
-    </ErrorBoundary>
-  )
-}
-
-const StyledMarket = styled.div`
-  width: 100%;
-  height: 90%;
-  position: absolute;
-  overflow-x: hidden;
-  overflow-y: auto !important;
-  &::-webkit-scrollbar {
-    width: 5px;
-    height: 15px;
-  }
-
-  &::-webkit-scrollbar-track-piece {
-    background-color: ${(props) => props.theme.color.grey[800]};
-  }
-
-  &::-webkit-scrollbar-thumb:vertical {
-    height: 30px;
-    background-color: ${(props) => props.theme.color.grey[700]};
-  }
-  scrollbar-color: ${(props) => props.theme.color.grey[700]}
-    ${(props) => props.theme.color.grey[800]};
-  scrollbar-width: thin;
-`
-
-export const StyledLitContainer = styled(Col)`
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-`
-
-const StyledHeaderContainer = styled.div`
-  //background-color: ${(props) => props.theme.color.grey[700]};
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  padding: ${(props) => props.theme.spacing[3]}px;
-`
-
-interface CardContainerProps {
-  multiplier: number
-}
-
-const StyledCardContainer = styled.span<CardContainerProps>`
-  background: ${(props) => props.theme.color.grey[800]};
-  //border: 1px solid grey;
-  border-radius: ${(props) => props.theme.borderRadius}px;
-  margin: ${(props) => props.theme.spacing[3]}px;
-  padding: ${(props) => props.theme.spacing[2]}px;
-  width: ${(props) => props.theme.contentWidth * (1 / props.multiplier)}px;
-  justify-content: center;
-  display: flex;
-  flex-direction: column;
-`
-
-const CardTitle = styled.span`
-  font-size: 18px;
-  color: ${(props) => props.theme.color.white};
-  opacity: 0.5;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  display: flex;
-  justify-content: center;
-`
-
-const Text = styled.span`
-  color: ${(props) => props.theme.color.white};
-  font-size: 24px;
-  font-weight: 700;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  display: flex;
-  justify-content: center;
-`
-
-const CardContent = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding: ${(props) => props.theme.spacing[2]}px;
-`
-const StyledContainer = styled(Col)`
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  flex-grow: 1;
-`
-
-export default Liquidity
